requirements.txt
appdirs==1.4.4
autopep8==1.5.4
black==20.8b1
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.17
django-allauth==0.43.0
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2020.1
regex==2020.11.13
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.26.2
whitenoise==5.2.0
